refactor: input field, button 분리
