+++ src/components/@common/DatePicker/index.tsx
@@ -1,7 +1,8 @@
 import {useState} from 'react';
 import DateTimePickerModal from 'react-native-modal-datetime-picker';
 import {getFormattedDate} from 'utils/formatDate';
-import InputForm from 'components/@common/InputForm';
+import InputField from 'components/@common/InputField';
+import {TouchableOpacity} from 'react-native-gesture-handler';
 
 type DatePickerProps = {date: Date; updateDate: (date: Date) => void};
 
@@ -22,20 +23,21 @@ export default function DatePicker({date, updateDate}: DatePickerProps) {
   };
 
   return (
-    <InputForm
-      label={'날짜'}
-      placeholder="날짜를 선택하세요"
-      onPress={handlePressDateTimePicker}
-      value={getFormattedDate(date)}
-      bottomSheet={
-        <DateTimePickerModal
-          isVisible={isDateTimePickerVisible}
-          mode="date"
-          date={date}
-          onConfirm={handleConfirmDateTimePicker}
-          onCancel={handleCancelDateTimePicker}
+    <>
+      <TouchableOpacity onPress={handlePressDateTimePicker}>
+        <InputField
+          label={'날짜'}
+          value={getFormattedDate(date)}
+          placeholder={'날짜를 선택하세요'}
         />
-      }
-    />
+      </TouchableOpacity>
+      <DateTimePickerModal
+        isVisible={isDateTimePickerVisible}
+        mode="date"
+        date={date}
+        onConfirm={handleConfirmDateTimePicker}
+        onCancel={handleCancelDateTimePicker}
+      />
+    </>
   );
 }
